requirements.txt
aniso8601==3.0.2
appdirs==1.4.3
appnope==0.1.0
astroid==2.2.4
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
coverage==4.5.3
coveralls==1.6.0
curtsies==0.3.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-heroku==0.3.1
djangorestframework==3.9.2
docopt==0.6.2
entrypoints==0.3
filelock==3.0.10
flake8==3.7.7
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.10
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.4
pbr==5.1.3
pep8==1.7.1
mock==2.0.0
parso==0.3.4
pbr==5.1.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
promise==2.2.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
toml==0.10.0
tox==3.8.1
traitlets==4.3.2
typed-ast==1.3.1
typing==3.6.6
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
whitenoise==4.1.2
wcwidth==0.1.7
wrapt==1.11.1
yapf==0.26.0
